# Perf optimization: reduce converter dictionary lookups