--- a/src/Nerdbank.MessagePack/Converters/ObjectMapConverter`1.cs
+++ b/src/Nerdbank.MessagePack/Converters/ObjectMapConverter`1.cs
@@ -104,7 +104,7 @@ internal class ObjectMapConverter<T>(MapSerializableProperties<T> serializable, 
 			int count = reader.ReadMapHeader();
 			for (int i = 0; i < count; i++)
 			{
-				ReadOnlySpan<byte> propertyName = CodeGenHelpers.ReadStringSpan(ref reader);
+				ReadOnlySpan<byte> propertyName = ReadStringSpan(ref reader);
 				if (deserializable.Value.Readers.TryGetValue(propertyName, out DeserializableProperty<T> propertyReader))
 				{
 					propertyReader.Read(ref value, ref reader, context);
@@ -155,7 +155,7 @@ internal class ObjectMapConverter<T>(MapSerializableProperties<T> serializable, 
 				int bufferedEntries = bufferedStructures / 2;
 				for (int i = 0; i < bufferedEntries; i++)
 				{
-					ReadOnlySpan<byte> propertyName = CodeGenHelpers.ReadStringSpan(ref syncReader);
+					ReadOnlySpan<byte> propertyName = ReadStringSpan(ref syncReader);
 					if (deserializable.Value.Readers.TryGetValue(propertyName, out DeserializableProperty<T> propertyReader))
 					{
 						propertyReader.Read(ref value, ref syncReader, context);
@@ -175,7 +175,7 @@ internal class ObjectMapConverter<T>(MapSerializableProperties<T> serializable, 
 					if (bufferedStructures % 2 == 1)
 					{
 						// The property name has already been buffered.
-						ReadOnlySpan<byte> propertyName = CodeGenHelpers.ReadStringSpan(ref syncReader);
+						ReadOnlySpan<byte> propertyName = ReadStringSpan(ref syncReader);
 						if (deserializable.Value.Readers.TryGetValue(propertyName, out DeserializableProperty<T> propertyReader) && propertyReader.PreferAsyncSerialization)
 						{
 							// The next property value is async, so turn in our sync reader and read it asynchronously.
@@ -206,5 +206,32 @@ internal class ObjectMapConverter<T>(MapSerializableProperties<T> serializable, 
 		}
 
 		return value;
+	}
+
+	/// <summary>
+	/// Reads a string as a contiguous span of UTF-8 encoded characters.
+	/// An array may be allocated if the string is not already contiguous in memory.
+	/// </summary>
+	/// <param name="reader">The reader to use.</param>
+	/// <returns>The span of UTF-8 encoded characters.</returns>
+	protected static ReadOnlySpan<byte> ReadStringSpan(scoped ref MessagePackReader reader)
+	{
+		if (!reader.TryReadStringSpan(out ReadOnlySpan<byte> result))
+		{
+			ReadOnlySequence<byte>? sequence = reader.ReadStringSequence();
+			if (sequence.HasValue)
+			{
+				if (sequence.Value.IsSingleSegment)
+				{
+					return sequence.Value.First.Span;
+				}
+
+				return sequence.Value.ToArray();
+			}
+
+			return default;
+		}
+
+		return result;
 	}
 }