… src/Nerdbank.MessagePack/Utilities/CodeGenHelpers.cs