Compare to #80 as baseline.
These tables were collected by different machines, so pay attention to the ratios not the means.

### Baseline

| Method                               | Categories             | Mean      | Error    | StdDev   | Median    | Ratio | RatioSD |
|------------------------------------- |----------------------- |----------:|---------:|---------:|----------:|------:|--------:|
| DeserializeAsArrayInit               | array-init,Deserialize |  59.57 ns | 0.409 ns | 0.362 ns |  59.63 ns |  0.91 |    0.01 |
| DeserializeAsArrayInit_MsgPackCSharp | array-init,Deserialize |  65.19 ns | 0.910 ns | 0.807 ns |  65.14 ns |  1.00 |    0.02 |
|                                      |                        |           |          |          |           |       |         |
| SerializeAsArrayInit                 | array-init,Serialize   |  84.25 ns | 0.775 ns | 0.725 ns |  84.19 ns |  1.05 |    0.01 |
| SerializeAsArrayInit_MsgPackCSharp   | array-init,Serialize   |  80.22 ns | 0.818 ns | 0.765 ns |  80.15 ns |  1.00 |    0.01 |
|                                      |                        |           |          |          |           |       |         |
| DeserializeAsArray                   | array,Deserialize      |  57.29 ns | 0.253 ns | 0.197 ns |  57.35 ns |  0.88 |    0.01 |
| DeserializeAsArray_MsgPackCSharp     | array,Deserialize      |  64.86 ns | 0.547 ns | 0.485 ns |  64.86 ns |  1.00 |    0.01 |
|                                      |                        |           |          |          |           |       |         |
| SerializeAsArray                     | array,Serialize        |  81.21 ns | 1.530 ns | 1.571 ns |  81.47 ns |  1.05 |    0.02 |
| SerializeAsArray_MsgPackCSharp       | array,Serialize        |  77.29 ns | 1.181 ns | 1.105 ns |  77.07 ns |  1.00 |    0.02 |
|                                      |                        |           |          |          |           |       |         |
| DeserializeMapInit                   | map-init,Deserialize   | 110.80 ns | 0.725 ns | 0.678 ns | 110.94 ns |  1.52 |    0.02 |
| DeserializeMapInit_MsgPackCSharp     | map-init,Deserialize   |  72.99 ns | 1.037 ns | 0.970 ns |  73.03 ns |  1.00 |    0.02 |
|                                      |                        |           |          |          |           |       |         |
| SerializeMapInit                     | map-init,Serialize     |  99.91 ns | 0.865 ns | 0.809 ns | 100.05 ns |  1.22 |    0.01 |
| SerializeMapInit_MsgPackCSharp       | map-init,Serialize     |  81.57 ns | 0.695 ns | 0.651 ns |  81.56 ns |  1.00 |    0.01 |
|                                      |                        |           |          |          |           |       |         |
| DeserializeMap                       | map,Deserialize        | 101.51 ns | 0.567 ns | 0.530 ns | 101.50 ns |  1.32 |    0.06 |
| DeserializeMap_MsgPackCSharp         | map,Deserialize        |  76.86 ns | 1.536 ns | 3.529 ns |  75.46 ns |  1.00 |    0.06 |
|                                      |                        |           |          |          |           |       |         |
| SerializeMap                         | map,Serialize          |  99.96 ns | 0.729 ns | 0.682 ns |  99.64 ns |  1.23 |    0.01 |
| SerializeMap_MsgPackCSharp           | map,Serialize          |  81.17 ns | 0.669 ns | 0.625 ns |  81.22 ns |  1.00 |    0.01 |

### With these changes

| Method                               | Categories             | Mean      | Error     | StdDev    | Median    | Ratio | RatioSD |
|------------------------------------- |----------------------- |----------:|----------:|----------:|----------:|------:|--------:|
| DeserializeAsArrayInit               | array-init,Deserialize |  94.61 ns |  4.176 ns | 12.182 ns |  94.75 ns |  0.76 |    0.16 |
| DeserializeAsArrayInit_MsgPackCSharp | array-init,Deserialize | 127.88 ns |  7.104 ns | 20.947 ns | 125.19 ns |  1.03 |    0.24 |
|                                      |                        |           |           |           |           |       |         |
| SerializeAsArrayInit                 | array-init,Serialize   | 145.20 ns |  8.137 ns | 23.991 ns | 136.01 ns |  0.89 |    0.23 |
| SerializeAsArrayInit_MsgPackCSharp   | array-init,Serialize   | 168.98 ns | 12.389 ns | 36.335 ns | 156.01 ns |  1.04 |    0.30 |
|                                      |                        |           |           |           |           |       |         |
| DeserializeAsArray                   | array,Deserialize      |  86.28 ns |  2.456 ns |  7.240 ns |  88.34 ns |  0.83 |    0.09 |
| DeserializeAsArray_MsgPackCSharp     | array,Deserialize      | 104.48 ns |  2.103 ns |  6.199 ns | 106.62 ns |  1.00 |    0.09 |
|                                      |                        |           |           |           |           |       |         |
| SerializeAsArray                     | array,Serialize        | 117.05 ns |  2.340 ns |  5.999 ns | 117.42 ns |  0.99 |    0.06 |
| SerializeAsArray_MsgPackCSharp       | array,Serialize        | 117.82 ns |  2.322 ns |  2.764 ns | 117.63 ns |  1.00 |    0.03 |
|                                      |                        |           |           |           |           |       |         |
| DeserializeMapInit                   | map-init,Deserialize   | 141.43 ns |  2.823 ns |  4.558 ns | 141.82 ns |  1.14 |    0.04 |
| DeserializeMapInit_MsgPackCSharp     | map-init,Deserialize   | 124.59 ns |  2.445 ns |  2.402 ns | 124.52 ns |  1.00 |    0.03 |
|                                      |                        |           |           |           |           |       |         |
| SerializeMapInit                     | map-init,Serialize     | 149.87 ns |  2.984 ns |  4.556 ns | 149.72 ns |  1.20 |    0.04 |
| SerializeMapInit_MsgPackCSharp       | map-init,Serialize     | 124.89 ns |  2.427 ns |  2.492 ns | 125.34 ns |  1.00 |    0.03 |
|                                      |                        |           |           |           |           |       |         |
| DeserializeMap                       | map,Deserialize        | 121.09 ns |  2.204 ns |  1.841 ns | 121.95 ns |  1.01 |    0.02 |
| DeserializeMap_MsgPackCSharp         | map,Deserialize        | 119.92 ns |  1.669 ns |  1.561 ns | 119.92 ns |  1.00 |    0.02 |
|                                      |                        |           |           |           |           |       |         |
| SerializeMap                         | map,Serialize          | 143.80 ns |  2.702 ns |  2.891 ns | 143.97 ns |  1.15 |    0.03 |
| SerializeMap_MsgPackCSharp           | map,Serialize          | 124.55 ns |  2.320 ns |  2.279 ns | 124.85 ns |  1.00 |    0.03 |